Fixes to Taverna Lite code to get DataTables working
Taverna Lite uses a modified portal along with an additional taverna_lite Gem.
Both the portal and the Gem include jquery_ujs, and this seems to cause an
error, which prevent DataTable from being included correctly.  In an attempt
to sort this, some reordering changes were made in the Git repo, but these
just make things worse on a new install.  So, we revert those changes and
remove one of the jquery_ujs includes, and all seems to work now.