Set "known good" portal version
Rather than use the latest version, we set a particular version for the
portal code. Updating to newer versions can then be tested before being
committed. This ensures we always have a "known good" configuration, if
problems are introduced in later versions, especially conflicts between
Ansible and the portal code, which cannot be picked up by developers not
using Ansible.