Prevent character encoding problems in bundle install
Running "bundle install" through ssh to Ubuntu seems to default to a
ASCII character encoding, which cause problems when Ruby gemspecs have
non-ASCII characters, e.g. in author names. Override the setting to
ensure it works.

Also, minor change to Solr config, so errors do not need to be ignored.